refactor(library/Base64): renames "index.ts" to "exports.ts"
